Improved.
Improved for large influx of people.

+++ main.py
@@ -2,7 +2,8 @@ import telebot
 from api import generate_image, save_image
 import os
 from config import BOT_TOKEN, LOG
-bot = telebot.TeleBot(BOT_TOKEN)
+from threading import Thread
+bot = telebot.TeleBot(BOT_TOKEN, threaded=True)
 
 @bot.message_handler(commands=['start'])
 def start(message):
@@ -10,22 +11,29 @@ def start(message):
 
 @bot.message_handler(content_types=['text'])
 def generate(message):
+    Thread(target=generate_image_thread, args=(message,)).start()
+
+def generate_image_thread(message):
     prompt = message.text
     if LOG:
         print(f"#{message.from_user.id}@{message.from_user.username}: {prompt}")
     bot.send_chat_action(message.chat.id, 'typing')
     bot.send_message(message.chat.id, 'Generating your image...')
-    try:
-        image_url = generate_image(prompt)
-        image_path = os.path.join(os.getcwd(), 'image.jpg')
-        save_image(image_url, image_path)
-        bot.send_chat_action(message.chat.id, 'upload_photo')
-        with open(image_path, 'rb') as photo:
-            bot.send_photo(message.chat.id, photo)
-        os.remove(image_path)
-        bot.send_message(message.chat.id, 'Your image has been generated!')
-    except:
+    
+    image_url = generate_image(prompt)
+    if image_url is None:
+        if LOG:
+            print(f'Error occurred while processing the prompt from user #{message.from_user.id}@{message.from_user.username}. The error could be due to NSFW content or an issue with the API.')
         bot.send_message(message.chat.id, 'Something went wrong, possibly, that was an NSFW alert, or problem with API, try again please.')
+        return
+    
+    image_path = os.path.join(os.getcwd(), 'image.jpg')
+    save_image(image_url, image_path)
+    bot.send_chat_action(message.chat.id, 'upload_photo')
+    with open(image_path, 'rb') as photo:
+        bot.send_photo(message.chat.id, photo)
+    os.remove(image_path)
+    bot.send_message(message.chat.id, 'Your image has been generated!')
 
 bot.polling(non_stop=True)
